+++ java/src/com/sleepycat/db/CacheFileStats.java
@@ -4,7 +4,7 @@
  *
  * See the file LICENSE for redistribution information.
  *
- * Copyright (c) 2002,2008 Oracle.  All rights reserved.
+ * Copyright (c) 2002, 2010 Oracle and/or its affiliates.  All rights reserved.
  */
 
 package com.sleepycat.db;
@@ -40,43 +40,43 @@ public class CacheFileStats {
         return st_map;
     }
 
-    private int st_cache_hit;
+    private long st_cache_hit;
     /**
     Requested pages found in the cache.
     */
-    public int getCacheHit() {
+    public long getCacheHit() {
         return st_cache_hit;
     }
 
-    private int st_cache_miss;
+    private long st_cache_miss;
     /**
     Requested pages not found in the cache.
     */
-    public int getCacheMiss() {
+    public long getCacheMiss() {
         return st_cache_miss;
     }
 
-    private int st_page_create;
+    private long st_page_create;
     /**
     Pages created in the cache.
     */
-    public int getPageCreate() {
+    public long getPageCreate() {
         return st_page_create;
     }
 
-    private int st_page_in;
+    private long st_page_in;
     /**
     Pages read into the cache.
     */
-    public int getPageIn() {
+    public long getPageIn() {
         return st_page_in;
     }
 
-    private int st_page_out;
+    private long st_page_out;
     /**
     Pages written from the cache to the backing file.
     */
-    public int getPageOut() {
+    public long getPageOut() {
         return st_page_out;
     }
